requirements.txt
xgboost
pandas
numpy
scikit-learn
isort
black
pycodestyle
matplotlib
python-dotenv